+++ AutomatedLabTest/internal/tests/ADFSProxy.tests.ps1
@@ -1,15 +1,8 @@
-﻿param
-(
-    [Parameter(Mandatory)]
-    [AutomatedLab.Lab]
-    $Lab
-)
-
-Describe "$($Lab.Name) ADFSProxy" -Tag ADFSProxy {
-
-    Context "Role deployment successful" {
-        It "Should return the correct amount of machines" {
-            (Get-LabVm -Role ADFSProxy).Count | Should -Be $Lab.Machines.Where({$_.Roles.Name -contains 'ADFSProxy'}).Count
+Describe "[$($(Get-Lab).Name)] ADFSProxy" -Tag ADFSProxy {
+Context "Role deployment successful" {
+        It "[ADFSProxy] Should return the correct amount of machines" {
+            (Get-LabVm -Role ADFSProxy).Count | Should -Be $(Get-Lab).Machines.Where({$_.Roles.Name -contains 'ADFSProxy'}).Count
         }
     }
 }
+