+++ AutomatedLabTest/internal/tests/FirstChildDC.tests.ps1
@@ -1,15 +1,8 @@
-﻿param
-(
-    [Parameter(Mandatory)]
-    [AutomatedLab.Lab]
-    $Lab
-)
-
-Describe "$($Lab.Name) FirstChildDC" -Tag FirstChildDC {
-
-    Context "Role deployment successful" {
-        It "Should return the correct amount of machines" {
-            (Get-LabVm -Role FirstChildDC).Count | Should -Be $Lab.Machines.Where({$_.Roles.Name -contains 'FirstChildDC'}).Count
+Describe "[$($(Get-Lab).Name)] FirstChildDC" -Tag FirstChildDC {
+Context "Role deployment successful" {
+        It "[FirstChildDC] Should return the correct amount of machines" {
+            (Get-LabVm -Role FirstChildDC).Count | Should -Be $(Get-Lab).Machines.Where({$_.Roles.Name -contains 'FirstChildDC'}).Count
         }
     }
 }
+