+++ AutomatedLabTest/internal/tests/SQLServer2012.tests.ps1
@@ -1,15 +1,8 @@
-﻿param
-(
-    [Parameter(Mandatory)]
-    [AutomatedLab.Lab]
-    $Lab
-)
-
-Describe "$($Lab.Name) SQLServer2012" -Tag SQLServer2012 {
-
-    Context "Role deployment successful" {
-        It "Should return the correct amount of machines" {
-            (Get-LabVm -Role SQLServer2012).Count | Should -Be $Lab.Machines.Where({$_.Roles.Name -contains 'SQLServer2012'}).Count
+Describe "[$($(Get-Lab).Name)] SQLServer2012" -Tag SQLServer2012 {
+Context "Role deployment successful" {
+        It "[SQLServer2012] Should return the correct amount of machines" {
+            (Get-LabVm -Role SQLServer2012).Count | Should -Be $(Get-Lab).Machines.Where({$_.Roles.Name -contains 'SQLServer2012'}).Count
         }
     }
 }
+